DataStore context API
Provide a "single source of truth" for data via the React Context API.
Wrap UI components in a <DataStore> component.
Any part of the UI that needs access to that data store can be placed in a <DataView> child component.

TODO: add DataStore props to define what/how data is retreived/stored.